+++ CHANGELOG.md
@@ -7,8 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+- [#766] `decoder::log`: Rename `PrettyLogger` to `StdoutLogger`
 - [#764]: Use mermaid for "defmt ecosystem" diagram
 
+[#766]: https://github.com/knurling-rs/defmt/pull/766
 [#764]: https://github.com/knurling-rs/defmt/pull/764
 
 ## [v0.3.5] - 2023-06-19